Remove remaining dependencies from core to contrib (in v2 only).
This is likely to break things in 2.0, only tests will tell. I fixed some of the things that will definitely break (saved_model stuff), but this definitely removes some kernels that used to be linked into some of the tools (specifically, graph_transforms), which we may have to put back.

PiperOrigin-RevId: 226946051