Refactor WaitForVariableChange: abstract away the atomic operations from it by putting them behind a `condition` function that it waits for, rename it `WaitUntil`, move it to its own library with own unit-test.
A design benefit from this change is that now the atomic operations are
grouped on the caller's side, which makes it easier to verify that they
are correct (i.e. that 'release' stores match 'acquire' loads).

Another design benefit is that this WaitUntil function now handles all of the spin-waiting that we do in ruy, including the one in BlockingCounter which was using separate code. That code had its own challenge as, not depending on a condition variable, it was occasionally reverting to just waiting for fixed durations to give other threads a chance to get scheduled, which could harm performance. That concern is removed by this change making it use WaitUntil. Making WaitUntil take a std::function for the wait condition was needed to make it general enough for this purpose: the original usage in ThreadPool needs to wait until a value is *different* from a given one, but the usage in BlockingCounter needs to wait until a value is *equal* to a given one.

An overload of WaitUntil allows controlling the spin_duration, in preparation for future changes exposing this setting in the ruy API itself.

PiperOrigin-RevId: 254184689