Fix for the test interpreter_test.py with PY3
The test interpreter_test.py is failing when it is run using PY3.
There are several issues due to difference between PY2 and PY3:

The string in PY2 is stored as bytes by default, but as unicode in PY3 by default.
In PY3 if the string is in bytes, it has the prefix 'b'.
We convert all strings to be the same.

There is an exception chaining in PY3.
As result, the message thrown from the exception is different.

There is a segmentation fault on Python3, when all tests in
the file interpreter_test are run. The bug is that the global variable
stores the pointer to the function, which goes away from the scope.