TFLM: Fix issue in the memory planner.
This issue applies to all models with variable tensors. The root cause is that the memory planner data gets overwritten upon copying variables. This can be fixed by moving assign variable tensor value after all the offsets are set.

With this CL, we can reduce the arena size for a few models, but let's leave them there for now as I'll have another CL which increases the required arena size.

PiperOrigin-RevId: 278726966
Change-Id: Ide76d091683641ef78c7ac287457f00d2fd2fc7f