add edge support in arduino project
This is an ugly solution - I had a hard time determining how to manipulate the makefile and 'transform arduino source' python script to allow separation of implementations. I am eager to have input from someone who would know how to better oranize this.

(but in a strictly technical sense this works and does not break support for the Nano 33 BLE Sense)