Depthwise convolution 3x3 per-channel int8 for dot-product ARM (4).
Build framework for testing per-channel routines.

Mostly a neutral change. It does slip in code for merging per-channel offsets. The merge is to take the existing constant offset and create a vector of offsets by combining with a vector of adjustments (clamping the range). The code for filling the adjustments comes in a later change.

PiperOrigin-RevId: 292916491
Change-Id: Iac8ed42ac24db493dd4b49702b0660079e461b7e