TFLM: Apply     -fno-threadsafe-statics   to all targets.
It has been enabled on bluepill & Hexagon, it should be enabled widely to reduce discrepancy.

From GCC doc:
-fno-threadsafe-statics

Do not emit the extra code to use the routines specified in the C++ ABI for thread-safe initialization of local statics. You can use this option to reduce code size slightly in code that doesn’t need to be thread-safe.

https://gcc.gnu.org/onlinedocs/gcc/C_002b_002b-Dialect-Options.html#index-fno-threadsafe-statics

TFLM is not designed to run in multi-threading environment so this should be safe and potentially reduce binary size.

PiperOrigin-RevId: 302820909
Change-Id: Idd87d4a45a1ea376ccf4eed17bedd56793146d78