Change SimpleMemoryAllocator to allow allocating at head of the area.
Previously tensors were assigned a memory address without the allocator being aware. This resulted in objects allocated in the persistent area (tail) to silently alias objects allocated in the tensor/scratch area. There is no allocation error, and no sanitizer can catch this error (as all data is in one runtime array).

The concept of child allocators locking parents is removed, as AllocateFromHead for the memory-planned tensors must be done before the planner's temporary (child) allocator is freed.

Also simplify the allocation code by tracking the current head and tail of the free area, and adjusting after allocations. Note: no need to remember original head/tail, as the arena is not owned by the allocator, and we're never freeing objects.

PiperOrigin-RevId: 306882659
Change-Id: I004bb56d2a1fe8166ea93c9176ec7434e91ecad7