Only run bazel locally (not as part of presubmits).
This is because the bazel build is already part of the presubmits and
the goal here is to enable developers to catch errors sooner locally.

Also, the fact that the micro test_all.sh script runs within a docker
container means that initiating a bazel build on the CI system would
mean downloading bazel to the docker image.