Increase the number of targets that are built for Bluepill.
This will help provide some safeguard against issues like #43126

Note that the tests are only built, not run because we are currently
unable to use renode as part of out continuous integration system.

Also note that some of the core tests (such as micro_interpreter_test)
do not build for the bluepill target and have been explicitly excluded.

Finally, note that while the kernel_circular_buffer_test builds for
bluepill, it does not pass and so had to be excluded.