CMSIS-NN: Fix incorrect flags for non-DSP/MVE processors
Fixes #42951

CalculateOpData() was not invoked for target Arm Cortex-M
processors without DSP or MVE extension like Cortex-M3 or Cortex-M0.

The fix is to remove the usage of feature flags __ARM_FEATURE_DSP
and __ARM_FEATURE_MVE and replace them with an appropriate
'if' condition where applicable. It is also part of the general
work to move the processor feature flags completely to CMSIS-NN.

Two reinterpret_casts are replaced with static casts along the
area where the feature flag chagne is done.