Some cleanup of the Makefile
This is progress towards reducing duplicated flags.

While this change directly touches bluepill, hifimini and
cortex_m_gcc_generic, it also moves some flags into the common Makefile
so it may have an impact on other targets too.

Manually confirmed that the size and cycles of the keyword_benchmark
test are identical before and after this change for xtensa_hifimini.

Relying on CI for the other targets.