Alternate approach (to PR #43240) to fixing ethos-u build errors.
Taking this opportunity to remove variables with the same names but in
different namespaces. This is one of the reasons we are incrementally
moving to a flat namespace (https://abseil.io/tips/130).

Additionally, having the constants in a local scope follow the style of
local variables.

The fix (hopefully) for Ethos-U is coming from the fact that the array
dimensions are now coming from constexpr variables.