Fix the absl module not found error
ToT workarounds the absl module not found error by using $CMAKE_MODULE_PATH,
but this is still likely to fail due to the fact that $CMAKE_MODULE_PATH is
a semicolon-separated list of directories specifying a search path for CMake
modules. In fact, if we set more than one path to $CMAKE_MODULE_PATH, the absl
module not found error occurs. This patch solves the problem by give |absl_DIR|
the unique CMake module path in the source tree.