micro: delete gmock/gtest includes in copy of SPACE_TO_DEPTH
Delete the gmock and gtest includes in the currently inactive
copy of op SPACE_TO_DEPTH. There is an annoying difference
between internal and external versions of the code that causes
clang-formatting errors in CI checks. When the code is imported
internally, the gmock and gtest includes become:

    #include "/internal/path/to/gmock/gmock.h"

and then clang-format wants them to be in the lower include block
(i.e., next to "flatbuffers/flatbuffers.h").

The solution for now is to delete gmock and gtest includes when
making these operator copies.