TFL: Update detection_postprocess kernel
Changing std::partial_sort to std::stable_sort in the case where
std::partial_sort use the full range. This is the case preventing
bit-exactness between TFL and TFLM. Problem with std::partial_sort
is that the order of equal elements is not guaranteed to be
preserved. Potentially std::partial_sort should be totally
replaced but leaving that for now.
Next step would be to add a custom stable sort to TFLM, as
std::stable_sort uses heap.