Remove the redundant quantize op in the back to back quantize op pairs
This is usually happen when post-training quantizing the model which has custom
ops on the output. The existing modifying io lib will produce the back to back
quantize op pairs which will cause accuracy dropping.

This cl tries to identify the back to back quantize op pairs pattern and remove
the first quantize ops in the pairs.

PiperOrigin-RevId: 359452170
Change-Id: Ib169208078c0b92ee56d9eef7cc7ac178c9f81a3