First step towards prototyping v2 of TFLM integration with external IDEs.
With this change we can create an output directory containing all the
sources and headers (including third_party) needed to build the TFLM
static library.

See tools/ci_build/test_project_generation.sh for some sample commands.

A next step will be to add the sources for the examples as well.

Progress towards: #47413