micro: Show memory plan if preprocessor define set
Changes MicroAllocator to call GreedyMemoryPlanner::PrintMemoryPlan when
the TF_LITE_SHOW_MEMORY_USE preprocessor symbol is defined. The
motivation is to make it possible for developers to show the memory plan
without needing to modify TfLM source code.

TF_LITE_SHOW_MEMORY_USE is a new preprocessor symbol. It is not used
elsewhere.