Add package-private InterpreterImpl class, which implements the methods in the InterpreterApi interface.  Move the implementation of the InterpreterApi methods from Interpeter into InterpreterImpl, with Interpreter now inheriting those method implementations from InterpreterImpl rather than directly defining them.
This allows significantly simplifying the build targets, since the
implementation of the non-experimental APIs no longer depends on the
source files for the experimental APIs, so we no longer need a separate
build target for the factory.

More importantly, it allows us to construct an AAR file for the
stable targets that doesn't include the experimental classes,
in particular the public Interpreter class that defines the
experimental methods.

PiperOrigin-RevId: 384693364
Change-Id: I1fbd084e8a9e76e1f9a849948ca87ffecede17af