(lite/java) Separate out NativeInterpreterWrapper parts that depend on experimental APIs.
Split some parts of NativeInterpreterWrapper which depend on experimental APIs
into a separate derived class NativeInterpreterWrapperExperimental,
and don't include that derived class in the "stable" build targets.

In particular NativeInterpreterWrapper.resetVariableTensors depends on
the C++ Interpreter::ResetVariableTensors method, which is experimental,
so we want to avoid depending on that from the "stable" build targets.

(Note that parts which depend on Interpreter::ModifyGraphWithDelegate
have NOT been moved to NativeInterpreterWrapperExperimental. The tentative
plan for those is to instead refactor them to use InterpreterBuilder::AddDelegate, and to promote that API to stable, in subsequent CLs.)

PiperOrigin-RevId: 385662997
Change-Id: I8a3e0ef9e573b75b9429facda0fc4fb2e92a869a