Add an On-Device Training CoLab example
The following CoLab illustrates the high-level flow of preparing a TensorFlow
Lite model, converting it to TensorFlow Lite model and running in TensorFlow
Lite runtime for a training use case.

The implementation is based on the Keras classification example
(https://www.tensorflow.org/tutorials/keras/classification) in the TensorFlow
official guide page.

PiperOrigin-RevId: 390260696
Change-Id: I49e96726ff119a6607d8e84aa03716f5e0d2b6f4