Add bias adjustments when the scale for the bias would cause overflow.
In full-integer quantization, bias_scale is determined as input_scale * filter_scale. When the two dependent scales are too small, the INT32 quantized bias values will overflow and break the model. This change is to introduce a logic implemented in TOCO quantizer to MLIR's QuantizationDriver.

Deleted temporary patch that used fixed threshold for this adjustment.

PiperOrigin-RevId: 392189936
Change-Id: I01489c587a554f13e76c16b9ef8fb393cfd8498d