Turn `Tensor` into an interface and move implementation into `TensorImpl`
Similar to `InterpreterApi`, this extracts an interface from the `Tensor` class that can be used for an alternative implementation. One difference is that the constructors for `Tensor` are not public, so we can turn it directly into an interface instead of having to extract the interface to a new class.

Also, remove the experimental warning from the `asReadOnlyBuffer()` method.

PiperOrigin-RevId: 394498811
Change-Id: Ief2f28d07686670c3f11c211632d5ad5834ca72b