Refactoring to avoid dependency on ModifyGraphWithDelegate from useXNNPACK.
Change NativeInterpreterWrapper's useXNNPACK method so that instead of
directly applying the XNNPACK delegate using ModifyGraphWithDelegate,
it just constructs the delegate and returns it (and thus
also rename the method as CreateXNNPACKDelegate), and instead have the
caller add that delegate to the list of delegates that will get applied.

This required adding a Java type for XNNPack Delegate.

PiperOrigin-RevId: 396591601
Change-Id: Iaeee9c96eaff3e1457ad292cba61fa4c9f991555