Use InterpreterBuilder::AddDelegate (attempt #2).
Use the non-experimental method InterpreterBuilder::AddDelegate to
add delegates before constructing the interpreter, rather
than using the experimental method Interpreter::ModifyGraphWithDelegate
to set delegates after the interpreter is already constructed.

A significant drawback of this refactoring is that if any delegates
are enabled, we end up having to construct the interpreter twice:
once without delegates, in order to determine whether the model
contains any unresolved flex ops in order to decide whether to
enable the flex delegate, and then again with all the delegates.

PiperOrigin-RevId: 401003838
Change-Id: I8fdf8a4728a7a98677c5f8cea895447df234310b